FeatureNet/requirements.txt
evalutils==0.2.4
scikit-learn>=0.21
scipy>=1.2.1
tqdm
opencv-python==4.5.1.48
torchvision>=0.10.0
torch>=1.6.0
vit-pytorch==1.2.4
Pillow==8.4.0
packaging==21.3
pytorch-pretrained-vit==0.0.7
scikit-image==0.19.3
PyWavelets
